moves join pairs helper methods to own class
