+++ sql/src/main/java/io/crate/analyze/relations/JoinPair.java
@@ -25,15 +25,10 @@ package io.crate.analyze.relations;
 import com.google.common.base.Function;
 import com.google.common.base.Objects;
 import io.crate.analyze.symbol.Symbol;
-import io.crate.operation.operator.AndOperator;
 import io.crate.planner.node.dql.join.JoinType;
 import io.crate.sql.tree.QualifiedName;
 
 import javax.annotation.Nullable;
-import java.util.Arrays;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
 
 public class JoinPair {
 
@@ -92,11 +87,15 @@ public class JoinPair {
         return Objects.hashCode(left, right, joinType, condition);
     }
 
-    private boolean equalsNames(QualifiedName left, QualifiedName right) {
+    public void replaceCondition(Function<? super Symbol, Symbol> replaceFunction) {
+        condition = replaceFunction.apply(condition);
+    }
+
+    boolean equalsNames(QualifiedName left, QualifiedName right) {
         return this.left.equals(left) && this.right.equals(right);
     }
 
-    private void replaceNames(QualifiedName left, QualifiedName right, QualifiedName newName) {
+    void replaceNames(QualifiedName left, QualifiedName right, QualifiedName newName) {
         if (this.left.equals(left) || this.left.equals(right)) {
             this.left = newName;
         }
@@ -105,117 +104,15 @@ public class JoinPair {
         }
     }
 
-    public void replaceCondition(Function<? super Symbol, Symbol> replaceFunction) {
-        condition = replaceFunction.apply(condition);
-    }
-
-    /**
-     * Find and return a {@link JoinPair} for the given relation names, also check for reversed names.
-     * If the {@link JoinType} is INNER and a pair is found for the reversed names, merge the join conditions.
-     */
-    public static JoinPair ofRelationsWithMergedConditions(QualifiedName left,
-                                                           QualifiedName right,
-                                                           List<JoinPair> joinPairs) {
-        JoinPair joinPair = JoinPair.ofRelations(left, right, joinPairs, true);
-        if (joinPair == null) {
-            // default to cross join (or inner, doesn't matter)
-            joinPair = new JoinPair(left, right, JoinType.CROSS);
-        }
-        // if it's an INNER, lets check for reverse relations and merge conditions
-        if (joinPair.joinType() == JoinType.INNER) {
-            JoinPair joinPairReverse = JoinPair.ofRelations(right, left, joinPairs, false);
-            if (joinPairReverse != null) {
-                joinPair.condition(AndOperator.join(Arrays.asList(joinPair.condition(), joinPairReverse.condition())));
+    boolean isOuterRelation(QualifiedName name) {
+        if (joinType.isOuter()) {
+            if (left.equals(name) && (joinType == JoinType.RIGHT || joinType == JoinType.FULL)) {
+                return true;
             }
-        }
-
-        return joinPair;
-    }
-
-    /**
-     * Returns the {@link JoinPair} of a pair of relation names based on the given list of join pairs.
-     * If <p>checkReversePairs</p> is true, also check for any {@link JoinPair} with switched names.
-     */
-    @Nullable
-    public static JoinPair ofRelations(QualifiedName left,
-                                       QualifiedName right,
-                                       List<JoinPair> joinPairs,
-                                       boolean checkReversePair) {
-        for (JoinPair joinPair : joinPairs) {
-            if (joinPair.equalsNames(left, right)) {
-                return joinPair;
-            }
-        }
-        // check if relations were switched due to some optimization
-        if (checkReversePair) {
-            for (JoinPair joinPair : joinPairs) {
-                if (joinPair.equalsNames(right, left)) {
-                    JoinPair reverseJoinPair = new JoinPair(
-                        joinPair.right,
-                        joinPair.left,
-                        joinPair.joinType().invert(),
-                        joinPair.condition);
-                    joinPairs.add(reverseJoinPair);
-                    return reverseJoinPair;
-                }
-            }
-        }
-
-        return null;
-    }
-
-    /**
-     * Returns all relation names which are part of an outer join.
-     */
-    public static Set<QualifiedName> outerJoinRelations(List<JoinPair> joinPairs) {
-        Set<QualifiedName> outerJoinRelations = new HashSet<>();
-        for (JoinPair joinPair : joinPairs) {
-            if (joinPair.joinType().isOuter()) {
-                outerJoinRelations.add(joinPair.left);
-                outerJoinRelations.add(joinPair.right);
-            }
-        }
-        return outerJoinRelations;
-    }
-
-    /**
-     * Rewrite names of matching join pair relations and inside the condition function
-     */
-    public static void rewriteNames(QualifiedName left,
-                                    QualifiedName right,
-                                    QualifiedName newName,
-                                    Function<? super Symbol, Symbol> replaceFunction,
-                                    List<JoinPair> joinPairs) {
-        for (JoinPair joinPair : joinPairs) {
-            joinPair.replaceNames(left, right, newName);
-            joinPair.replaceCondition(replaceFunction);
-        }
-    }
-
-    /**
-     * Returns true if relation name is part of an outer join and on the outer side.
-     */
-    static boolean isOuterRelation(QualifiedName name, List<JoinPair> joinPairs) {
-        for (JoinPair joinPair : joinPairs) {
-            if (isOuterRelation(name, joinPair)) {
+            if (right.equals(name) && (joinType == JoinType.LEFT || joinType == JoinType.FULL)) {
                 return true;
             }
         }
         return false;
     }
-
-    public static boolean isOuterRelation(QualifiedName name, JoinPair joinPair) {
-        if (joinPair.joinType().isOuter()) {
-            if (joinPair.left.equals(name) &&
-                (joinPair.joinType() == JoinType.RIGHT || joinPair.joinType() == JoinType.FULL)) {
-                return true;
-            }
-            if (joinPair.right.equals(name) &&
-                (joinPair.joinType() == JoinType.LEFT || joinPair.joinType() == JoinType.FULL)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
 }
